requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.8
flask-wtf==1.1.1
beautifulsoup4==4.12.2
lxml==4.9.3
flask-assets==2.0